ref：<技能系统>重构技能CD部分，利用行为树适配各种复杂的CD需求

+++ Server/Model/NKGMOBA/Battle/SkillSystem/Others/SkillDesNodeData.cs
@@ -39,7 +39,7 @@ namespace ETModel
         [HideLabel]
         [Title("技能CD", Bold = false)]
         [BsonDictionaryOptions(DictionaryRepresentation.ArrayOfArrays)]
-        public Dictionary<int, float> SkillCD;
+        public Dictionary<int, long> SkillCD;
         
         [TabGroup("基础信息")]
         [HideLabel]
